refactor: rename routes to appointment and contact
